+++ udata/tests/api/test_datasets_api.py
@@ -38,7 +38,7 @@ from udata.core.dataset.models import (
 )
 from udata.core.organization.factories import OrganizationFactory
 from udata.core.spatial.factories import GeoLevelFactory, SpatialCoverageFactory
-from udata.core.topic.factories import TopicFactory
+from udata.core.topic.factories import TopicElementDatasetFactory, TopicFactory
 from udata.core.user.factories import AdminFactory, UserFactory
 from udata.i18n import gettext as _
 from udata.models import CommunityResource, Dataset, Follow, Member, db
@@ -182,7 +182,8 @@ class DatasetAPITest(APITestCase):
         format_dataset = DatasetFactory(resources=[ResourceFactory(format="my-format")])
         featured_dataset = DatasetFactory(featured=True)
         topic_dataset = DatasetFactory()
-        topic = TopicFactory(datasets=[topic_dataset])
+        topic = TopicFactory()
+        TopicElementDatasetFactory(element=topic_dataset, topic=topic)
 
         paca, _, _ = create_geozones_fixtures()
         geozone_dataset = DatasetFactory(spatial=SpatialCoverageFactory(zones=[paca.id]))
@@ -2426,7 +2427,7 @@ class HarvestMetadataAPITest:
         date = datetime(2022, 2, 22, tzinfo=pytz.UTC)
 
         harvest_metadata = HarvestResourceMetadata(
-            created_at=date,
+            issued_at=date,
             modified_at=date,
             uri="http://domain.gouv.fr/dataset/uri",
         )
@@ -2435,32 +2436,46 @@ class HarvestMetadataAPITest:
         response = api.get(url_for("api.dataset", dataset=dataset))
         assert200(response)
         assert response.json["resources"][0]["harvest"] == {
-            "created_at": date.isoformat(),
+            "issued_at": date.isoformat(),
             "modified_at": date.isoformat(),
             "uri": "http://domain.gouv.fr/dataset/uri",
         }
 
     def test_dataset_with_harvest_computed_dates(self, api):
-        creation_date = datetime(2022, 2, 22, tzinfo=pytz.UTC)
+        # issued_date takes precedence over internal creation date and harvest created_at on dataset
+        issued_date = datetime(2022, 2, 22, tzinfo=pytz.UTC)
+        creation_date = datetime(2022, 2, 23, tzinfo=pytz.UTC)
+        modification_date = datetime(2022, 3, 19, tzinfo=pytz.UTC)
+        harvest_metadata = HarvestDatasetMetadata(
+            created_at=creation_date,
+            issued_at=issued_date,
+            modified_at=modification_date,
+        )
+        dataset = DatasetFactory(harvest=harvest_metadata)
+        response = api.get(url_for("api.dataset", dataset=dataset))
+        assert200(response)
+        assert response.json["created_at"] == issued_date.isoformat()
+        assert response.json["last_modified"] == modification_date.isoformat()
+
+        # without issuance date, creation_date takes precedence over internal creation date on dataset
         modification_date = datetime(2022, 3, 19, tzinfo=pytz.UTC)
         harvest_metadata = HarvestDatasetMetadata(
             created_at=creation_date,
             modified_at=modification_date,
         )
         dataset = DatasetFactory(harvest=harvest_metadata)
-
         response = api.get(url_for("api.dataset", dataset=dataset))
         assert200(response)
         assert response.json["created_at"] == creation_date.isoformat()
         assert response.json["last_modified"] == modification_date.isoformat()
 
+        # issued_date takes precedence over internal creation date on resource
         resource_harvest_metadata = HarvestResourceMetadata(
-            created_at=creation_date,
+            issued_at=issued_date,
             modified_at=modification_date,
         )
         dataset = DatasetFactory(resources=[ResourceFactory(harvest=resource_harvest_metadata)])
-
         response = api.get(url_for("api.dataset", dataset=dataset))
         assert200(response)
-        assert response.json["resources"][0]["created_at"] == creation_date.isoformat()
+        assert response.json["resources"][0]["created_at"] == issued_date.isoformat()
         assert response.json["resources"][0]["last_modified"] == modification_date.isoformat()